+++ Sources/ShopifyAcceleratedCheckouts/Wallets/AcceleratedCheckoutButtons.swift
@@ -38,10 +38,10 @@ public enum RenderState {
 /// - The `wallets` modifier can be used to limit the buttons rendered
 /// - The order of the buttons is the same as the order of the `wallets` modifier
 /// - omission of the `wallets` modifier will render all buttons
-@available(iOS 17.0, *)
+@available(iOS 16.0, *)
 public struct AcceleratedCheckoutButtons: View {
-    @Environment(ShopifyAcceleratedCheckouts.Configuration.self)
-    private var configuration
+    @EnvironmentObject
+    private var configuration: ShopifyAcceleratedCheckouts.Configuration
 
     let identifier: CheckoutIdentifier
     var wallets: [Wallet] = [.shopPay, .applePay]
@@ -49,10 +49,14 @@ public struct AcceleratedCheckoutButtons: View {
     var cornerRadius: CGFloat?
 
     /// The Apple Pay button label style
-    private var applePayLabel: ApplePayButtonLabel = .plain
+    private var applePayLabel: PayWithApplePayButtonLabel = .plain
 
     @State private var shopSettings: ShopSettings?
-    @State private var currentRenderState: RenderState = .loading
+    @State private var currentRenderState: RenderState = .loading {
+        didSet {
+            eventHandlers.renderStateDidChange?(currentRenderState)
+        }
+    }
 
     /// Initializes an Apple Pay button with a cart ID
     /// - Parameters:
@@ -98,13 +102,10 @@ public struct AcceleratedCheckoutButtons: View {
                             )
                         }
                     }
-                }.environment(shopSettings)
+                }.environmentObject(shopSettings)
             }
         }
         .task { await loadShopSettings() }
-        .onChange(of: currentRenderState) { _, newValue in
-            eventHandlers.renderStateDidChange?(newValue)
-        }
         .onAppear {
             eventHandlers.renderStateDidChange?(currentRenderState)
         }
@@ -131,9 +132,9 @@ public struct AcceleratedCheckoutButtons: View {
 
 // MARK: AcceleratedCheckoutButtons Modifiers
 
-@available(iOS 17.0, *)
+@available(iOS 16.0, *)
 extension AcceleratedCheckoutButtons {
-    public func applePayLabel(_ label: ApplePayButtonLabel) -> AcceleratedCheckoutButtons {
+    public func applePayLabel(_ label: PayWithApplePayButtonLabel) -> AcceleratedCheckoutButtons {
         var view = self
         view.applePayLabel = label
         return view