+++ Sources/ShopifyAcceleratedCheckouts/Wallets/ApplePay/ApplePayAuthorizationDelegate/ApplePayState.swift
@@ -24,7 +24,7 @@
 import Foundation
 import PassKit
 
-@available(iOS 17.0, *)
+@available(iOS 16.0, *)
 enum ApplePayState: Equatable {
     static func == (lhs: ApplePayState, rhs: ApplePayState) -> Bool {
         return String(describing: lhs.self) == String(describing: rhs.self)