+++ Sources/ShopifyAcceleratedCheckouts/Wallets/ApplePay/Data/PassKitFactory.swift
@@ -23,7 +23,7 @@
 
 import PassKit
 
-@available(iOS 17.0, *)
+@available(iOS 16.0, *)
 class PassKitFactory {
     static let shared = PassKitFactory()